+++ src/main/java/conekta/io/model/submodel/DiscountLine.java
@@ -3,12 +3,11 @@ package conekta.io.model.submodel;
 import conekta.io.model.ConektaObject;
 import lombok.Data;
 
-import java.math.BigDecimal;
 
 @Data
 public class DiscountLine extends ConektaObject {
     private String code;
-    private BigDecimal amount;
+    private Integer amount;
     private String type;
     private String parentId;
     private Boolean deleted;
